+++ lib/libesp32/berry_tasmota/src/be_udp_lib.cpp
@@ -14,9 +14,14 @@
 // extern int be_udp_begin_mcast(bvm *vm);
 
 #include <Arduino.h>
-#include <WiFiGeneric.h>
 #include <WiFiUdp.h>
 #include "be_mapping.h"
+
+// Tasmota Logging
+extern void AddLog(uint32_t loglevel, PGM_P formatP, ...);
+enum LoggingLevels {LOG_LEVEL_NONE, LOG_LEVEL_ERROR, LOG_LEVEL_INFO, LOG_LEVEL_DEBUG, LOG_LEVEL_DEBUG_MORE};
+
+extern bool WifiHostByName(const char* aHostname, IPAddress & aResult);
 
 extern "C" {
 
@@ -37,11 +42,11 @@ extern "C" {
     return be_call_c_func(vm, (void*) &be_udp_deinit_ntv, "=.p", "");
   }
 
-  // udp.begin(address:string, port:int) -> bool
+  // udp.begin(interface:string, port:int) -> bool
   int32_t be_udp_begin_ntv(WiFiUDP *udp, const char *host, int32_t port) {
-    IPAddress addr((uint32_t)0);
-    // if no host or host is "" then we defult to INADDR_ANY (0.0.0.0)
-    if(host && (*host != 0) && !WiFiGenericClass::hostByName(host, addr)){
+    IPAddress addr;
+    // if no host or host is "" then we defult to INADDR_ANY
+    if(host && (*host != 0) && !WifiHostByName(host, addr)){
         return 0;
     }
     return udp->begin(addr, port);
@@ -60,8 +65,8 @@ extern "C" {
 
   // udp.begin_multicast(address:string, port:int) -> nil
   int32_t be_udp_begin_mcast_ntv(WiFiUDP *udp, const char *host, int32_t port) {
-    IPAddress addr((uint32_t)0);
-    if(!WiFiGenericClass::hostByName(host, addr)){
+    IPAddress addr;
+    if(!WifiHostByName(host, addr)){
         return 0;
     }
     return udp->WiFiUDP::beginMulticast(addr, port);
@@ -72,10 +77,11 @@ extern "C" {
 
   // udp.send(address:string, port:int, payload:bytes) -> bool
   int32_t be_udp_send_ntv(WiFiUDP *udp, const char *host, int32_t port, const uint8_t* buf, int32_t len) {
-    IPAddress addr((uint32_t)0);
-    if (!WiFiGenericClass::hostByName(host, addr)){
+    IPAddress addr;
+    if (!WifiHostByName(host, addr)){
         return 0;
     }
+    // AddLog(LOG_LEVEL_DEBUG, "BRY: udp.begin got host '%s'", addr.toString().c_str());
     if (!udp->beginPacket(addr, port)) { return 0; }
     int bw = udp->write(buf, len);
     if (!bw) { return 0; }